% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_plot.R
\name{contour_plot}
\alias{contour_plot}
\title{Contour plots of functions of two variables}
\usage{
contour_plot(
  object,
  formula,
  domain,
  npts = 100,
  labels = TRUE,
  filled = TRUE,
  contours_at = NULL,
  n_contours = 10,
  n_fill = 50,
  alpha = 1,
  fill_alpha = 0.1,
  label_alpha = 1,
  contour_color = NULL,
  label_color = ifelse(!is.null(contour_color), contour_color, "black"),
  skip = 1,
  guide = FALSE,
  guide_title = "output",
  color_scale = scale_color_viridis_c(),
  fill_scale = scale_fill_viridis_c(),
  ...
)
}
\description{
Contour plots of functions of two variables
}
