% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcalc_symbolic.R
\name{convert_R_to_Ryacas}
\alias{convert_R_to_Ryacas}
\title{Simplify a mathematical expression in a function using Ryacas}
\usage{
convert_R_to_Ryacas(X)
}
\arguments{
\item{X}{an R expression implementing a mathematical function}

\item{fun}{The function whose body contains a mathematical expression.}
}
\description{
Simplify a mathematical expression in a function using Ryacas
}
