% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid_US.R
\docType{data}
\name{Covid_US}
\alias{Covid_US}
\title{COVID data from the first half of the pandemic}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 189 rows and 18 columns.
}
\usage{
Covid_US
}
\description{
COVID data from the first half of the pandemic
}
\keyword{datasets}
