% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv-patterns.R
\name{differentiation_rules}
\alias{differentiation_rules}
\title{Rules for differentiation}
\arguments{
\item{X}{an R expression such as \code{a*x + b}}
}
\description{
Each \verb{_rule} function checks the input expression for a particular kind of structure.
If that structure is found, the function returns an expression for the \strong{derivative}.
But if the pattern sought is not found, the function returns FALSE.
}
\details{
Currently, these differentiate with respect to \code{x}
}
\examples{
x_rule(x)
axb_rule(x)
axb_rule(a*x + 7)
}
