% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PP.R
\name{flow_field}
\alias{flow_field}
\title{Functions for drawing phase-plane dynamics}
\usage{
flow_field(
  dynamics,
  xrange = 0:1,
  yrange = 0:1,
  n = 10,
  pseudo = FALSE,
  nsteps = 10,
  stepsize = 0.05
)
}
\arguments{
\item{dynamics}{a list containing two one-sided formulas specifying
the dynamics, e.g. \code{list(~ x + y, ~ y - x)}}

\item{n}{number of arrows to draw on a side. $n^2$ will be the number of arrows
overall.}

\item{pseudo}{Logical flag. Whether to position the arrows pseudo-randomly rather than
on a strict grid.}
}
\description{
Functions for drawing phase-plane dynamics
}
