% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun-with-boxes.R
\name{graph_with_boxes}
\alias{graph_with_boxes}
\title{Draw a graph with labeled boxes}
\usage{
graph_with_boxes(
  fn,
  intervals = tibble::tibble(x = -1:1, xend = x + 0.6),
  domain = c(-3, 3),
  my_letters = LETTERS,
  ...
)
}
\description{
Draw a graph with labeled boxes
}
