% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_and_tail.R
\name{head_and_tail}
\alias{head_and_tail}
\title{Display head and tail of a long data frame}
\usage{
head_and_tail(df, n = 4L, ntop = n, nbottom = n, nmiddle = 0L)
}
\arguments{
\item{df}{A data frame or data-frame-like object}

\item{n}{A number specifying the number of rows to show at the top and bottom.}

\item{ntop}{Overrides \code{n} to specify the number of rows to show at the top}

\item{nbottom}{Overrides \code{n} to specify the number of rows at the bottom}

\item{nmiddle}{Integer. Whether to show  intermediate row to indicate
that there are elided rows}
}
\description{
Works just like \code{head()} or \code{tail} but shows both the first and last rows
}
\examples{
head_and_tail(mtcars)
mtcars \%>\% head_and_tail(ntop=2, nbottom=4)
mtcars \%>\% head_and_tail(nmiddle = 3)

}
