% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.R
\name{iterate}
\alias{iterate}
\title{Iterate a function on an initial condition}
\usage{
iterate(f = cos, x0 = 0, n = 10, ...)
}
\arguments{
\item{f}{a function of one or more state variables which returns a vector containing
those same state variables. Parameters to \code{f()}, if any, should be named and at
the end of the argument list.}

\item{x0}{a vector with the numerical initial condition. There should be 1 component
in \code{x0} for each of the state variables in \code{f()}.}

\item{n}{an integer specifying the number of iterations}

\item{...}{values for numerical parameters to the function \code{f()}}
}
\value{
A data frame with a column \code{.i} listing the iteration number
and columns for each component of the initial condition. There will be n+1
rows, the first for the initial condition and the remaining for the n iterations.
}
\description{
Iterates a function a specified number of times on an initial condition.
}
\details{
The function \code{f} can take one or more arguments. The first of these
should represent the \emph{state} of the dynamical system, e.g. x or x and y, etc.
At the end of the argument list to \code{f} can come named parameters. The length of the
initial condition \code{x0} must match the number of state arguments. Numerical values for
parameters (if any) must be provided in the \code{...} slot.
}
\examples{
iterate(function(x, mu=3.5) mu*x*(1-x), .232, n=10, mu=4) # chaos
iterate(function(x, y) c(x+y, x-y), c(1,1), n=10)
iterate(function(x, y) c(x+y, x), c(0,1), n=10) # fibonacci
}
