% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ivan.R
\docType{data}
\name{Ivan}
\alias{Ivan}
\title{Track of Hurricane Ivan in Sept. 2004}
\format{An object of class \code{data.frame} with 95 rows and 23 columns.}
\usage{
Ivan
}
\description{
Data: https://www.nhc.noaa.gov/data/hurdat/hurdat2-1851-2019-042820.txt
Format for data: https://www.nhc.noaa.gov/data/hurdat/hurdat2-format-nov2019.pdf
}
\keyword{datasets}
