% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcalc_symbolic.R
\name{mcalc_simplify}
\alias{mcalc_simplify}
\title{Simplify a mathematical expression in a function using Ryacas}
\usage{
mcalc_simplify(fun)
}
\arguments{
\item{fun}{The function whose body contains a mathematical expression.}
}
\description{
Simplify a mathematical expression in a function using Ryacas
}
