% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prosperity.R
\docType{data}
\name{Properity}
\alias{Properity}
\alias{Prosperity}
\title{Measures of material well-being as a function of income.}
\format{A data frame  with 6 rows and 11 variables.
\itemize{
\item \code{CrimeProblem} - Fraction of families living in an area with a crime problem
\item \code{AbandonedBuildings} - Fraction of families living in an area with abandoned building
\item \code{IncompleteBathroom} - Fraction of families without a complete bathroom in their dwelling.
\item Other variables from \code{NoCentralHeat} to \code{Telephone} are similarly a fraction of families
\item \code{DoctorVisitsUnder7} - (?) Average number of visits to a doctor of children under 7
\item \code{DoctorVisits7To18} - (?) Average number of visits to a doctor of children in this age group.
\item \code{NoDoctorVisitUnder7} - (?) Fraction of households with children under 7 who have
not visited a doctor (ever? in the last year?)
\item \code{NoDoctorVisit7To18} - Similar, but for children aged 7 to 18.
}}
\source{
Susan E. Mayer (1997) "What money can’t buy: Family income and children’s life chances," Harvard Univ. Press p. 102.
}
\usage{
data(Prosperity)
}
\description{
The data divides Americans into six groups based on income. Within
each group, the other variables report the percentage of households with children in the indicated
situation. Doctor visits are counts.
}
\details{
The file ``Income-Housing.csv" contains data on incomes and housing conditions for US households divided into groups by income. (Source:
}
\keyword{datasets}
