% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_tide.R
\docType{data}
\name{RI_tide}
\alias{RI_tide}
\title{Tide level in Providence, Rhode Island US}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 6480 rows and 2 columns.}
\usage{
RI_tide
}
\description{
Minute-by-minute tide data during April 1-5 2010.
}
\details{
Lat/Long: 41.8071N, -71.4012W
Request: \verb{https://tidesandcurrents.noaa.gov/api/datagetter?begin_date=20200401\%2010:00&end_date=20200405\%201:59&station=8454000&product=one_minute_water_level&datum=mllw&units=metric&time_zone=gmt&application=web_services&format=csv}
Other product documentation: https://tidesandcurrents.noaa.gov/api/
Google Map: https://www.google.com/maps/place/Providence,+RI/@41.8071,-71.4012,14z/data=!4m5!3m4!1s0x89e444e0437e735d:0x69df7c4d48b3b627!8m2!3d41.8071!4d-71.4012
}
\keyword{datasets}
