% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roof_fun.R
\name{roof_fun}
\alias{roof_fun}
\title{A mathematical function shaped like a roof}
\usage{
roof_fun(x, y, type = c("plain", "dormer", "silo"))
}
\arguments{
\item{x}{the x-coordinates at which to evaluate the function}

\item{y}{the y-coordinates at which to evaluate the function}

\item{type}{one of \code{"plain"},  \code{"dormer"}, or \code{"silo"}.}
}
\description{
Roof-shaped functions to illustrate the properties of partial derivatives.
}
\details{
Note that when used in \code{contour_plot()} or \code{interactive_plot()}, you
need only specify the symbolic expression \code{roof_fun(x, y) ~ x + y}. The plotting function
will calculate and provide the needed numerical values for the coordinate.
The domain for this function is \code{x=c(-3,3)} and \code{y=c(-1,1)}.
}
\examples{
\dontrun{
interactive_plot(roof_fun(x, y, type="dormer") ~ x & y,  domain(x=c(-3,3), y=c(-1,1)))
}
}
