% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-axes.R
\name{scale_shift}
\alias{scale_shift}
\title{Draw parallel axes with a linear transform}
\usage{
scale_shift(min, max, r, x0 = 0, nticks = 10)
}
\description{
Only handles linear transformations: r*(x-x0)
See examples
}
