% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-dynamics.R
\name{show_matrix_dynamics}
\alias{show_matrix_dynamics}
\title{Draw diagram of the flow induced by a matrix}
\usage{
show_matrix_dynamics(A)
}
\description{
Draw diagram of the flow induced by a matrix
}
