% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PP.R
\name{streamlines}
\alias{streamlines}
\title{Functions for drawing phase-plane dynamics}
\usage{
streamlines(
  dynamics,
  domain = list(x = c(0, 1), y = c(0, 1)),
  n = 10,
  nsteps = 10,
  stepsize = 0.05
)
}
\arguments{
\item{dynamics}{a list containing two one-sided formulas specifying
the dynamics, e.g. \code{list(~ x + y, ~ y - x)}}

\item{domain}{list of two vectors specifying domain of the initial
points of each streamline Vector of length 2 specifying the domain}

\item{n}{number of arrows to draw on a side. $n^2$ will be the number of arrows
overall.}
}
\description{
Functions for drawing phase-plane dynamics
}
\examples{
streamlines(list(~ u+v, ~ u - v), stepsize=0.01, n=20)

}
