% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traj_plot.R
\name{traj_plot}
\alias{traj_plot}
\title{Plots to correspond with integrateODE()}
\usage{
traj_plot(one, two, soln, npts = 500, nt = 5, domain = NULL, ...)
}
\arguments{
\item{one}{a formula describing the plot, e.g. B(t) ~ R(t) or B(t) ~ t or a ggplot
object}

\item{two}{placeholder for handling when \code{one} is a ggplot object}

\item{soln}{the solution as produced by \code{integrateODE()}}

\item{npts}{number of plotted points (default: 500)}

\item{nt}{number of tick marks to use in a trajectory plot}

\item{domain}{Optional list like \code{domain=domain(t=c(0,100))}. By default, this
will be inferred from \code{soln}}
}
\description{
Utilities to work with solutions to ODEs as generated by integrateODE
}
